refactor(schema): move LocalHoodieSchemaCache
Move LocalHoodieSchemaCache from common.util to common.schema, next to
the other schema caches (HoodieSchemaCache, HoodieAvroSchemaCache), and
rename its factory from getInstance() to create(): it returns a new,
empty cache on every call, so the old name suggested a singleton it is
not. RecordContext is the only caller.

Note: RecordContext is Serializable and holds the cache as a
non-transient field, so the Java-serialized form now names the new
package. RecordContext is per-task state and is never persisted.

Part of #16639.
