+++ hudi-common/src/main/java/org/apache/hudi/common/engine/RecordContext.java
@@ -24,11 +24,11 @@ import org.apache.hudi.common.model.DeleteRecord;
 import org.apache.hudi.common.model.HoodieOperation;
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.schema.HoodieSchema;
+import org.apache.hudi.common.schema.LocalHoodieSchemaCache;
 import org.apache.hudi.common.table.HoodieTableConfig;
 import org.apache.hudi.common.table.read.BufferedRecord;
 import org.apache.hudi.common.table.read.DeleteContext;
 import org.apache.hudi.common.util.JavaTypeConverter;
-import org.apache.hudi.common.util.LocalHoodieSchemaCache;
 import org.apache.hudi.common.util.OrderingValues;
 import org.apache.hudi.common.util.collection.ArrayComparable;
 import org.apache.hudi.common.util.collection.Pair;
@@ -65,7 +65,7 @@ public abstract class RecordContext<T> implements Serializable {
 
   private final SerializableBiFunction<T, HoodieSchema, String> recordKeyExtractor;
   // for encoding and decoding schemas to the spillable map
-  private final LocalHoodieSchemaCache localSchemaCache = LocalHoodieSchemaCache.getInstance();
+  private final LocalHoodieSchemaCache localSchemaCache = LocalHoodieSchemaCache.create();
 
   @Getter
   protected final JavaTypeConverter typeConverter;
